#49, the app builds with many mistakes

+++ app/src/main/java/com/jumparoundcreations/sugarcounter/data/historyData/GraphData.kt
@@ -2,7 +2,7 @@ package com.jumparoundcreations.sugarcounter.data.historyData
 
 data class GraphData(
     val id: Int,
-    val valueTotal: Int,
+    val valueTotal: Double,
     val day: String,
     val date: String
 )
